Renamed and added size limit
