+++ main.dogo
@@ -4,10 +4,10 @@ import (
 	"net/http"
 	"os"
 
-	"github.com/ob-algdatii-20ss/leistungsnachweis-teammaze/api"
+	"github.com/ob-algdatii-20ss/leistungsnachweis-teammaze/web"
 )
 
 func main() {
-	http.Handle("/", api.MazeAPIRouter())
+	http.Handle("/", web.MazeAPIRouter())
 	_ = http.ListenAndServe(os.Getenv("ADDRESS"), nil)
 }
